Introduce v0.4.0 with type checking and enhanced documentation
This commit introduces version 0.4.0 and adds type checking to the `attributes` and `attribute` methods in Barley. An error will now be raised if the object's attribute is not of the specified type. Documentation has been updated to guide users on this new type-checking feature.

Additionally, a link to the API documentation has also been included to provide users with a more comprehensive understanding of Barley and its capabilities.